feat: Enhance memory analysis configuration and improve code readability
- Added new configuration entries for memory analysis in application.properties:
  - analysis.memory.heap-persistence.max-points
  - analysis.memory.heap-growth.max-points
  - analysis.memory.old-gen-persistence.max-points
  - analysis.memory.old-gen-growth.max-points
  - analysis.memory.gc-retention.max-points
  - analysis.memory.histogram-growth.max-points
  - analysis.cpu.top-threads

- Refactored ConfigLoader to include new analysis properties.
- Updated SnapshotHandler and SnapshotService for better readability.
- Changed leak confidence type from int to double in JvmHistorySample and JvmDeltaSnapshot.
- Improved HistoricalLeakAnalyzer to enhance clarity and maintainability.
- Introduced numeric helper methods in GrowthCalculator for clamping and conversions.

+++ src/main/java/com/acorp/jvminsight/snapshotcollection/dto/JvmHistorySample.java
@@ -20,7 +20,7 @@ public record JvmHistorySample(
     long processCpuTimeNanos,
     long processReadBytes,
     long processWriteBytes,
-    int leakConfidence) {
+    double leakConfidence) {
 
   public static JvmHistorySample from(JvmSnapshot snapshot) {
     long heap = snapshot.getMemory() == null ? 0 : snapshot.getMemory().heapUsed;
@@ -50,7 +50,7 @@ public record JvmHistorySample(
         snapshot.getProcessCpu() == null ? 0 : snapshot.getProcessCpu().processCpuTimeNanos();
     long readBytes = snapshot.getProcessIo() == null ? 0 : snapshot.getProcessIo().readBytes();
     long writeBytes = snapshot.getProcessIo() == null ? 0 : snapshot.getProcessIo().writeBytes();
-    int confidence = snapshot.getDelta() == null ? 0 : snapshot.getDelta().getLeakScore();
+    double confidence = snapshot.getDelta() == null ? 0 : snapshot.getDelta().getLeakScore();
 
     return new JvmHistorySample(
         snapshot.getTimestamp(),
